requirements.txt
fastapi
uvicorn
pymongo
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
email-validator
python-jose[cryptography]

